Fix track number to be 1 based, change reset timings, add pin confirmation
Track number on empeg serial port is zero-based, so make it 1-based when we interpret it.

Change around the reset timings to try to fix some pairing bugs.

Try to work around GitHub issue #2  by resetting the bluetooth if the arduino resets, and resetting the global variables if the bluetooth resets. Experimental.

Add PIN confirmation "OK" command to pairing process to help Mark Lord's connection issue which is GitHub issue #7 though this will probably not fix it completely.