V523 changes and ResetBluetoothPin routine (provisional)
This version contains a change in behavior to compensate for the deduplication change in Hijack V523. Instead of waiting for a "trigger" to flag the host stereo for new metadata, it flags the host stereo for every single new line of metadata received. This causes some problems, though, discussing with mark maybe going back to non-deduplication mode.

This version also contains a routine to use Pin 51 to physically reset the bluetooth module from its RST line, though it is currently disabled due to my board getting fried when I tried it.