Move explanatory text into a README.txt file (Step 1)
Initial work to move explanatory text into a README.txt file to fix issue #16. This commit adds the file, a later commit will remove the text from the source code and touch up the source code in various spots to correctly refer to the README.txt instead of saying "above".