Force Reset, timing, and other bugfixes.
This set of changes addresses issues #20 and #22 in partial ways. Issue #22 seems to be doing better but issue #20 can't be fixed, so it's closed now.

Also this update changes the code in the functions DisplayAndProcessCommands and DisplayAndSwallowResponses so that they are based on an actual timer instead of a "Delay(1)" command. Hoping that this will improve the serial buffer overflows in situations where it overflows during these functions running.