Partial fix to issue #25.
Partial temporary fix to issue #25, which is caused by a string in Arduino running out of memory. Temporarily fix the issue by passing the string as a reference instead of as a value. This fixes that instance of that string but now there are issues where the logging has some string mangling. More work needed on issue #25.

Begin experimenting with problems with EmpegStartPause as described in issue  #26.