Default to Yes/No prompt security.
This fixes issue #14. The BlueGigaEmpeg module now defaults to the security scheme where the head unit prompts for a particular number on the stereo and the bluetooth module must respond with a yes/no answer. The code will always respond with the "yes" answer to this question.

This security scheme was required by Mark Lord's car stereo, it wouldn't work with "Just Works" pairing mode.