Do not send a CALL 17 AVRCP on the first A2DP connection.
Bugfix, related to but not solving issue #21, and maybe hopefully solving issue #29 (still need to test). When first pairing up with the host stereo, I was originally jumping in and doing a CALL (addr) 17 AVRCP as soon as I got the first CONNECT A2DP message. This is a problem because that means it's trying to set up an AVRCP connection while the host stereo is in the middle of setting up its second (stereo) A2DP channel. This is a problem because it causes security errors during the pairing process. The resulting connection seems to work OK but there was an error message in the process. So fix that by only asking for an AVRCP connection once the SECOND A2DP connection gets registered.