Fix issue #26 - EmpegStartPause
Fix the issue #26 - "EmpegStartPause does not work as expected".

There were several other cases and situations where the BlueGigaEmpeg module should be telling the empeg to be paused. Plugged as many of those holes as I could.

Now that this is complete, the next task is to fix issue #42 - Detect when bluetooth is connected. Then we could do away with the EmpegStartPause flag altogether.