Candidate fix for issue #45
Don't do as many unneccesary "A2DP Streaming Start" commands. These might be confusing and inducing bugs in the host stereo's bluetooth operating system. One of the first things that happens after initial connection is that the host stereo sends a play command to the player, and it's possible that the extra "streaming start" is causing undefined behavior at that crucial time.

This is just a theory that needs to be borne out over time/testing.