Update codec commands because codecs were being weird.
There were problems where codecs weren't working right and I saw it was setting more than one codec to priority 0 so I 'm trying to make it forget all codecs except the one I want. That isn't working so I'm trying to set the other codecs to lower prioroties. Not understanding how this works, the dox are bad on this topic. Possibly related to issue #45 but I'm not certain.