Remove ResetLine code and fix issue #42 .
New design with 3v3 constant power does not require the reset line pin on the Arduino, remove the corresponding code.

Fix issue #42, create a flag to detect when bluetooth is connected and streaming, and perform different behaviors depending on whether or not it is streaming. First behavior is to pause the player at bootup only if bluetooth is not detected as being paried+streaming. Other behaviors may come later.