Add Monkey reconnect and TypeZToPair
Partial fixes for issue #60 but they need some bake time to be sure.

Add a different reconnect method, Monkey Reconnect, which uses my own code to reconnect instead of the automatic ones built into iWrap6.

Add feature TypeZToPair which will, if enabled, go into pair mode if you type a Z on the console. Helpful for debugging if the pair button is in the boot.

A good bug fix for pair mode: Turns out it never clears the prior address if it's currently connected. Fix that.

Some other minor tweaks and bug fixes, code comment cleanup.