Implement Shuffle, attempt to fix issue #61
Implemented a response to "GET APPLICATION SETTING ATTRIBUTES" which is simultaneously an attempt to fix issue #61 and also happens to attempt to implement shuffle (since I had to respond with something, I chose Shuffle as the thing to respond with).

Does not fix issue $61 but I now think that's not fixable on my and.

Can't test Shuffle just now.