Continued work on reconnect feature
Reorganized notes on the reconnect feature and the various things I tried in an attempt to fix issue #71 "No AVRCP on Onkyo".

This includes many comment updates listing everything wrong with all other possible configurations of the reconnect command.

Need to test this current version on Onkyo to see if it fixes issue #71.