Add configurable auto-reconnect modes.
This contains a work-around for issue #71 "Onkyo AVRCP Commands Do Not Work". If a user encounters a problem where the device does not reconnect to AVRCP, then they can set this to mode "2" which uses an alternate reconnection string.

There is also an outside ghost of a chance that the newest default reconnection string might also fix the Onkyo problem. But even if it does I want the alternate string option in the code anyway.