Boot chip in analog mode then switch to digital
Attempt to fix issue #70 (Broken up high pitched playback on cold boot) by booting the module into analog mode, then changing it to digitan and doing another reset. This happens in two places: At boot, and at pairtime.

The issue is so rare that I need to just check this in and live with it a while to see if it fixes it.