Experimental changes to pairing and discoverability.
Attempts to find a happy medium betweeen issues #28 and #37 by making the BGE less discoverable when it has an existing pairing buddy, and more discoverable when it doesn't have a pairing buddy yet, or when you have pressed the RESET/PAIR button.

Make the pairing system alternate between 5 seconds of searching and 5 seconds of listening during the pairing procedure. This hopefully solves issue #37 without breaking anything else, which allows me to pair in the Honda either when I have pressed the pairing button or when I have not pressed the pairing button.

Update code comments in several places.

Small changes related to issue #70 and #67 (chipmunk playback or silent playback), though neither of those issues are fully fixed yet.