Add intelligent streaming start commands.
Use more intelligence for the correct time to issue a Streaming Start command. This was an attempt to improve the behavior around issues #67 (silent playback) and issue #70 (48khz cracking chipmunk playback). I believe that it should theoretically completely fix issue #67, though it hasn't put a dent in issue #70 (though it didn't make it any worse).

There are some other minor code tweaks and improvements in this version as well.