Version 1.1.10 - Revert streaming start behavior back to old "dumb" method.
The "smart" streaming start code turns out to induce issue #70 more frequently instead of decreasing it.

Reverting back to prior method of issuing streaming starts, which is a "dumb" method based on the scFixMessageMatrix table.

This will re-induce issue #67 but that issue was very rare so I will live with it until I can figure out what exactly is going on with these streaming start commands.