1.1.13 - New streaming start behavior. (#82)
New version 1.1.13. Change the behavior of Streaming Start commands to look for the second and third receipt of any RING or CONNECT and issue a streaming start on each one. I expect this will fix issue #67 completely. It hasn't really put a dent in recurrences of issue #70, but that one is handled successfully by the reboot workaround.